Distributed Services into multiple class corresponding to their entities.
